PSR https://www.php-fig.org/psr/
> PSR-1
3. Namespace and Class Names
- Class names MUST be declared in StudlyCaps.

> PSR-2
2.2. Files
- All PHP files MUST end with a single blank line.
- The closing tag MUST be omitted from files containing only PHP.
2.4. Indenting
- Code MUST use an indent of 4 spaces, and MUST NOT use tabs for indenting.
4.3. Methods
- Method names MUST NOT be declared with a space after the method name. The
opening brace MUST go on its own line, and the closing brace MUST go on the
next line following the body. There MUST NOT be a space after the opening
parenthesis, and there MUST NOT be a space before the closing parenthesis.
4.4. Method Arguments
- In the argument list, there MUST NOT be a space before each comma, and there
MUST be one space after each comma.
- Argument lists MAY be split across multiple lines, where each subsequent line
is indented once. When doing so, the first item in the list MUST be on the
next line, and there MUST be only one argument per line.
- When the argument list is split across multiple lines, the closing parenthesis
and opening brace MUST be placed together on their own line with one space
between them.
4.6. Method and Function Calls
- When making a method or function call, there MUST NOT be a space between the
method or function name and the opening parenthesis, there MUST NOT be a space
after the opening parenthesis, and there MUST NOT be a space before the
closing parenthesis. In the argument list, there MUST NOT be a space before
each comma, and there MUST be one space after each comma.